Effects Filters styling and cleanup
