+++ Client/src/components/admin/pages/inventory/effects/EffectListingPage.tsx
@@ -1,10 +1,9 @@
-import { Content } from 'rsuite';
+import { Content, Panel } from 'rsuite';
 import EffectListing from '../../../listing/EffectListing';
 import useTitle from '../../../../../hooks/useTitle';
 import AdminHeader from '../../../../common/header/AdminHeader';
 import { IEffectFilters } from '../../../../../types/IFilter';
 import { useEffect, useState } from 'react';
-import { debounce } from '../../../../../helpers/timing';
 import axios from '../../../../../api/axios';
 import EffectFilters from '../../../filters/EffectFilters';
 
@@ -33,45 +32,23 @@ const EffectListingPage = () => {
         getFilterData();
     }, [draw]);
 
-    const setFilterByKey = (key: keyof IEffectFilters, value: string | number) => {
-        setFilters({ ...filters, [key]: value });
-    };
-
-    const setValueRange = debounce((range: [number, number]) => {
-        setFilters({ ...filters, vmin: range[0], vmax: range[1] });
-    }, 500);
-
-    const setDurationRange = debounce((range: [number, number]) => {
-        setFilters({ ...filters, dmin: range[0], dmax: range[1] });
-    }, 500);
-
-    const clearFilters = () => {
-        setDraw(draw + 1);
-        setFilters({
-            name: '',
-            vmin: filterLimits.vmin ?? 0,
-            vmax: filterLimits.vmax ?? 1000,
-            dmin: filterLimits.dmin ?? 0,
-            dmax: filterLimits.dmax ?? 1000,
-        });
-    };
-
     if (loading) return <>Loading Screen</>;
 
     return (
-        <div className='admin-page'>
+        <Panel className='admin-page'>
             <AdminHeader title='Effects' />
             <Content>
                 <EffectFilters
+                    filters={filters}
                     filterLimits={filterLimits}
-                    setFilterByKey={setFilterByKey}
-                    setValueRange={setValueRange}
-                    setDurationRange={setDurationRange}
-                    clearFilters={clearFilters}
+                    setFilters={setFilters}
+                    onClearCallback={() => {
+                        setDraw(draw + 1);
+                    }}
                 />
                 <EffectListing filters={filters} />
             </Content>
-        </div>
+        </Panel>
     );
 };
