+++ Client/src/scss/components/_checkbox.scss
@@ -19,7 +19,7 @@
     .rs-checkbox-inner {
         &::before {
             // Checkbox box
-            background: $SeparatorColor;
+            background: $InkColor;
             border: 1px solid $SeparatorColor;
         }
     }
