+++ Client/src/scss/components/_filters.scss
@@ -51,6 +51,5 @@
         display: flex;
         justify-content: flex-end;
         gap: 10px;
-        padding-right: 10px;
     }
 }
