+++ Client/src/scss/pages/_admin-page.scss
@@ -11,7 +11,10 @@
 .admin-page {
     overflow-y: auto;
     max-height: 100vh;
-    padding: 10px;
+
+    .rs-panel-body {
+        padding: 10px;
+    }
 }
 
 .admin-sidebar {
